# packages/contracts/openapi.bundle.yaml
openapi: 3.1.0
info:
  title: Spott API
  version: 1.0.0
  description: Spott iOS/Web 单一契约。具备 Idempotency-Key 参数的副作用写入由调用方提供稳定幂等键； 媒体上传、完成、附件与举报均以调用方保留的 UUID 和服务端持久回执支持响应丢失恢复； 业务状态始终由服务端裁决。
  contact:
    name: Spott Engineering
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.spott.jp/v1
    description: Production
security:
  - BearerAuth: []
...
paths:
...
  /auth/email/verify:
    post:
      operationId: verifyEmailChallenge
      summary: Verify email challenge
      tags:
        - Auth
      security: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyChallengeRequest'
      responses:
        '200':
          $ref: '#/components/responses/AuthSession'
        '400':
          $ref: '#/components/responses/Error'
...
  /auth/refresh:
    post:
      operationId: refreshSession
      summary: Refresh session
      description: Rotates the current refresh credential. Idempotency-Key and a persistent device binding proof are optional for compatibility only while the supplied credential is still current and being used for the first time; neither omission permits recovery of a consumed credential.
      tags:
        - Auth
      security: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: Stable UUID for an independently retryable refresh attempt.
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
                - deviceId
              properties:
                refreshToken:
                  type: string
                deviceId:
                  $ref: '#/components/schemas/UUID'
                deviceBindingProof:
                  $ref: '#/components/schemas/PersistentDeviceBindingProof'
      responses:
        '200':
          $ref: '#/components/responses/AuthSession'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
...
  /sessions/{id}:
    delete:
      operationId: revokeSession
      summary: Revoke session
      tags:
        - Auth
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204':
          description: Revoked
        '404':
          $ref: '#/components/responses/Error'
...
  /groups:
    get:
      operationId: discoverGroups
      summary: Discover public groups by region, category, or query
      tags:
        - Groups
      security: []
      parameters:
        - name: region
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
            maxLength: 120
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Group discovery page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupPage'
        '400':
          $ref: '#/components/responses/Error'
    post:
      operationId: createGroup
      summary: Create group
      tags:
        - Groups
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupInput'
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          $ref: '#/components/responses/Error'
...
  /ops/auth/email/verify:
    post:
      operationId: verifyOpsEmailChallenge
      summary: Verify an operator email OTP and establish an MFA-gated Ops cookie session
      tags:
        - Ops
        - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyChallengeRequest'
      responses:
        '200':
          description: Sets __Host-spott_ops and __Host-spott_ops_refresh HttpOnly Secure SameSite cookies
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpsAuthSession'
        '403':
          $ref: '#/components/responses/Error'
  /ops/auth/refresh:
    post:
      operationId: refreshOpsSession
      summary: Rotate an Ops cookie session using the HttpOnly refresh cookie
      tags:
        - Ops
        - Auth
      security: []
      responses:
        '200':
          description: Rotated Ops cookies
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpsAuthSession'
        '401':
          $ref: '#/components/responses/Error'
  /ops/auth/session:
    delete:
      operationId: revokeOpsSession
      summary: Revoke the current Ops session and clear both cookies
      tags:
        - Ops
        - Auth
      security:
        - OpsSession: []
        - BearerAuth: []
      responses:
        '204':
          description: Session revoked
        '403':
          $ref: '#/components/responses/Error'
  /ops/overview:
    get:
      operationId: getOpsOverview
      summary: Get scoped queue, health, and growth totals
      tags:
        - Ops
      security:
        - OpsSession: []
        - BearerAuth: []
      responses:
        '200':
          description: Operations overview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpsOverview'
        '403':
          $ref: '#/components/responses/Error'
  /ops/users:
    get:
      operationId: getOpsUsers
      summary: Search redacted user accounts without email or phone values
      tags:
        - Ops
      security:
        - OpsSession: []
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: restriction
          in: query
          schema:
            type: string
        - name: deviceRisk
          in: query
          schema:
            type: string
            enum:
              - normal
              - elevated
              - blocked
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Redacted users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpsUserPage'
        '403':
          $ref: '#/components/responses/Error'
  /ops/users/{id}/restriction-decisions:
    post:
      operationId: decideOpsUserRestriction
      summary: Update account restrictions or submit a sensitive login-block decision for second approval
      tags:
        - Ops
      security:
        - OpsSession: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpsUserRestrictionInput'
      responses:
        '200':
          description: Applied or pending-approval restriction decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpsUserRestrictionDecision'
        '403':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /ops/organizers:
    get:
      operationId: getOpsOrganizers
      summary: List organizer quality and safety aggregates
      tags:
        - Ops
      security:
        - OpsSession: []
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Organizer aggregates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpsOrganizerPage'
        '403':
          $ref: '#/components/responses/Error'
  /ops/events:
    get:
      operationId: getOpsEvents
      summary: List submitted events with organizer and risk context
      tags:
        - Ops
        - Events
      security:
        - OpsSession: []
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: riskMin
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 100
        - name: region
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Scoped event review queue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpsEventPage'
        '403':
          $ref: '#/components/responses/Error'
  /ops/events/{id}/review:
    post:
      operationId: reviewEvent
      summary: Publish, return, or reject a submitted event with optimistic locking
      tags:
        - Ops
        - Events
      security:
        - OpsSession: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - decision
                - reason
              properties:
                decision:
                  type: string
                  enum:
                    - published
                    - needs_changes
                    - rejected
                reason:
                  type: string
                  minLength: 3
                  maxLength: 2000
      responses:
        '200':
          description: Review result; posterJobId is set when an enabled event is published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpsEventReviewResult'
        '403':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /ops/groups:
    get:
      operationId: getOpsGroups
      summary: List groups with lifecycle, transfer, report, and supply aggregates
      tags:
        - Ops
        - Groups
      security:
        - OpsSession: []
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Scoped group inventory
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpsGroupPage'
        '403':
          $ref: '#/components/responses/Error'
  /ops/groups/{id}/lifecycle-decision:
    post:
      operationId: decideOpsGroupLifecycle
      summary: Restore, close, cancel closure, or remove a group
      tags:
        - Ops
        - Groups
      security:
        - OpsSession: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - decision
                - reason
              properties:
                decision:
                  type: string
                  enum:
                    - restore
                    - start_closing
                    - cancel_closing
                    - remove
                reason:
                  type: string
                  minLength: 3
                  maxLength: 2000
      responses:
        '200':
          description: Lifecycle decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpsGroupLifecycleDecision'
        '409':
          $ref: '#/components/responses/Error'
  /ops/moderation/cases:
    get:
      operationId: getModerationCases
      summary: Get the redacted moderation queue with filters
      tags:
        - Ops
      security:
        - OpsSession: []
        - BearerAuth: []
      parameters:
        - name: severity
          in: query
          schema:
            type: string
            enum:
              - p0
              - p1
              - p2
        - name: status
          in: query
          schema:
            type: string
        - name: assignee
          in: query
          schema:
            $ref: '#/components/schemas/UUID'
        - name: targetType
          in: query
          schema:
            type: string
            enum:
              - event
              - group
              - user
              - comment
              - announcement
        - name: q
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Redacted moderation queue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpsModerationCasePage'
        '403':
          $ref: '#/components/responses/Error'
  /ops/moderation/cases/{id}:
    get:
      operationId: getModerationCase
      summary: Get redacted case details; purpose grants an audited short-lived evidence URL
      tags:
        - Ops
      security:
        - OpsSession: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - name: purpose
          in: query
          schema:
            type: string
            minLength: 3
            maxLength: 1000
      responses:
        '200':
          description: Redacted moderation details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpsModerationCaseDetail'
        '404':
          $ref: '#/components/responses/Error'
  /ops/moderation/cases/{id}/claim:
    post:
      operationId: claimModerationCase
      summary: Claim a moderation case with optimistic locking
      tags:
        - Ops
      security:
        - OpsSession: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: Claimed case
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpsModerationDecision'
        '409':
          $ref: '#/components/responses/Error'
  /ops/moderation/cases/{id}/decision:
    post:
      operationId: decideModerationCase
      summary: Decide a moderation case and append immutable action, audit, and outbox records
      tags:
        - Ops
      security:
        - OpsSession: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - decision
                - reason
              properties:
                decision:
                  type: string
                  enum:
                    - no_action
                    - hide
                    - remove
                    - restrict
                reason:
                  type: string
                  minLength: 3
                  maxLength: 2000
                durationHours:
                  type: integer
                  minimum: 1
                  maximum: 8760
      responses:
        '200':
          description: Immutable moderation decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpsModerationDecision'
        '409':
          $ref: '#/components/responses/Error'
...
  /ops/config-revisions/{id}/impact-preview:
    post:
      operationId: previewConfigRevisionImpact
      summary: Preview affected users, regions, warnings, and quote protection
      tags:
        - Ops
      security:
        - OpsSession: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Impact preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpsConfigImpact'
        '403':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    OpsSession:
      type: apiKey
      in: cookie
      name: __Host-spott_ops
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        oneOf:
          - $ref: '#/components/schemas/UUID'
          - type: string
            maxLength: 100
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/UUID'
...
  responses:
    Error:
      description: Stable business error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    AuthSession:
      description: Rotated device-bound session
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthSession'
  schemas:
    UUID:
      type: string
      format: uuid
...
    VerifyChallengeRequest:
      type: object
      required:
        - challengeId
        - code
        - deviceId
      properties:
        challengeId:
          $ref: '#/components/schemas/UUID'
        code:
          type: string
          pattern: ^[0-9]{6}$
        deviceId:
          $ref: '#/components/schemas/UUID'
...